Model/requirements.txt
# Core ML & Data
numpy
pandas
scikit-learn
scipy

# NLP / Embeddings
sentence-transformers
torch

# Geolocation
geopy

# Backend / API
fastapi
uvicorn

# Data validation
pydantic

# Utilities
python-multipart

# Environment variables
python-dotenv
requests
gradio
